Move shared types to CodeMirrorBlob
