+++ client/web/src/repo/blob/Blob.tsx
@@ -4,7 +4,7 @@ import classNames from 'classnames'
 import { Remote } from 'comlink'
 import * as H from 'history'
 import { isEqual } from 'lodash'
-import { Location, createPath, NavigateFunction, useLocation, useNavigate } from 'react-router-dom-v5-compat'
+import { createPath, useLocation, useNavigate } from 'react-router-dom-v5-compat'
 import {
     BehaviorSubject,
     combineLatest,
@@ -43,17 +43,12 @@ import { ActionItemAction } from '@sourcegraph/shared/src/actions/ActionItem'
 import { FlatExtensionHostAPI } from '@sourcegraph/shared/src/api/contract'
 import { haveInitialExtensionsLoaded } from '@sourcegraph/shared/src/api/features'
 import { ViewerId } from '@sourcegraph/shared/src/api/viewerTypes'
-import { ExtensionsControllerProps } from '@sourcegraph/shared/src/extensions/controller'
 import { getHoverActions } from '@sourcegraph/shared/src/hover/actions'
 import { HoverContext, PinOptions } from '@sourcegraph/shared/src/hover/HoverOverlay'
 import { getModeFromPath } from '@sourcegraph/shared/src/languages'
-import { PlatformContextProps } from '@sourcegraph/shared/src/platform/context'
-import { Settings, SettingsCascadeProps } from '@sourcegraph/shared/src/settings/settings'
-import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
-import { ThemeProps } from '@sourcegraph/shared/src/theme'
+import { Settings } from '@sourcegraph/shared/src/settings/settings'
 import { codeCopiedEvent } from '@sourcegraph/shared/src/tracking/event-log-creators'
 import {
-    AbsoluteRepoFile,
     FileSpec,
     ModeSpec,
     UIPositionSpec,
@@ -67,86 +62,11 @@ import { Code, useObservable } from '@sourcegraph/wildcard'
 
 import { getHover, getDocumentHighlights } from '../../backend/features'
 import { WebHoverOverlay } from '../../components/shared'
-import { BlobStencilFields, ExternalLinkFields, Scalars } from '../../graphql-operations'
-import { BlameHunkData } from '../blame/useBlameHunks'
-import { HoverThresholdProps } from '../RepoContainer'
 
 import { BlameColumn } from './BlameColumn'
+import { BlobInfo, BlobProps, updateBrowserHistoryIfChanged } from './CodeMirrorBlob'
 
 import styles from './Blob.module.scss'
-
-// Logical grouping of props that are only used by the CodeMirror blob view
-// implementation.
-interface CodeMirrorBlobProps {
-    overrideBrowserSearchKeybinding?: boolean
-}
-
-export interface BlobProps
-    extends SettingsCascadeProps,
-        PlatformContextProps,
-        TelemetryProps,
-        HoverThresholdProps,
-        ExtensionsControllerProps,
-        ThemeProps,
-        CodeMirrorBlobProps {
-    className: string
-    wrapCode: boolean
-    /** The current text document to be rendered and provided to extensions */
-    blobInfo: BlobInfo
-    'data-testid'?: string
-
-    // When navigateToLineOnAnyClick=true, the code intel popover is disabled
-    // and clicking on any line should navigate to that specific line.
-    navigateToLineOnAnyClick?: boolean
-
-    // Enables experimental navigation by rendering links for all interactive tokens.
-    enableLinkDrivenCodeNavigation?: boolean
-    // Enables experimental navigation by making interactive tokens selectable on click.
-    enableSelectionDrivenCodeNavigation?: boolean
-
-    // If set, nav is called when a user clicks on a token highlighted by
-    // WebHoverOverlay
-    nav?: (url: string) => void
-    role?: string
-    ariaLabel?: string
-
-    supportsFindImplementations?: boolean
-
-    isBlameVisible?: boolean
-    blameHunks?: BlameHunkData
-
-    navigate: NavigateFunction
-
-    /**
-     * TODO(valery): RR6
-     * @deprecated prefer using useNavigate()
-     */
-    history: H.History
-    /**
-     * TODO(valery): RR6
-     * @deprecated prefer using useLocation()
-     */
-    location: Location
-}
-
-export interface BlobInfo extends AbsoluteRepoFile, ModeSpec {
-    /** The raw content of the blob. */
-    content: string
-
-    /** The trusted syntax-highlighted code as HTML */
-    html: string
-
-    /** LSIF syntax-highlighting data */
-    lsif?: string
-
-    stencil?: BlobStencilFields[]
-
-    /** If present, the file is stored in Git LFS (large file storage). */
-    lfs?: { byteSize: Scalars['BigInt'] } | null
-
-    /** External URLs for the file */
-    externalURLs?: ExternalLinkFields[]
-}
 
 const domFunctions = {
     getCodeElementFromTarget: (target: HTMLElement): HTMLTableCellElement | null => {
@@ -755,40 +675,7 @@ export const Blob: React.FunctionComponent<React.PropsWithChildren<BlobProps>> =
     )
 }
 
-/**
- * Adds an entry to the browser history only if new search parameters differ
- * from the current ones. This prevents adding a new entry when e.g. the user
- * clicks the same line multiple times.
- */
-export function updateBrowserHistoryIfChanged(
-    navigate: NavigateFunction,
-    location: H.Location,
-    newSearchParameters: URLSearchParams,
-    /** If set to true replace the current history entry instead of adding a new one. */
-    replace: boolean = false
-): void {
-    const currentSearchParameters = [...new URLSearchParams(location.search).entries()]
-
-    // Update history if the number of search params changes or if any parameter
-    // value changes. This will also work for file position changes, which are
-    // encoded as parameter without a value. The old file position will be a
-    // non-existing key in the new search parameters and thus return `null`
-    // (whereas it returns an empty string in the current search parameters).
-    const needsUpdate =
-        currentSearchParameters.length !== [...newSearchParameters.keys()].length ||
-        currentSearchParameters.some(([key, value]) => newSearchParameters.get(key) !== value)
-
-    if (needsUpdate) {
-        const entry = {
-            ...location,
-            search: formatSearchParameters(newSearchParameters),
-        }
-
-        navigate(entry, { replace })
-    }
-}
-
-export function getLSPTextDocumentPositionParameters(
+function getLSPTextDocumentPositionParameters(
     position: HoveredToken & RepoSpec & RevisionSpec & FileSpec & ResolvedRevisionSpec,
     mode: string
 ): RepoSpec & RevisionSpec & ResolvedRevisionSpec & FileSpec & UIPositionSpec & ModeSpec {
