+++ client/web/src/repo/blob/BlobPage.tsx
@@ -68,9 +68,9 @@ import { ToggleLineWrap } from './actions/ToggleLineWrap'
 import { ToggleRenderedFileMode } from './actions/ToggleRenderedFileMode'
 import { getModeFromURL } from './actions/utils'
 import { fetchBlob, fetchStencil } from './backend'
-import { Blob, BlobInfo } from './Blob'
+import { Blob } from './Blob'
 import { BlobLoadingSpinner } from './BlobLoadingSpinner'
-import { Blob as CodeMirrorBlob } from './CodeMirrorBlob'
+import { Blob as CodeMirrorBlob, type BlobInfo } from './CodeMirrorBlob'
 import { GoToRawAction } from './GoToRawAction'
 import { BlobPanel } from './panel/BlobPanel'
 import { RenderedFile } from './RenderedFile'
