+++ client/web/src/repo/blob/codemirror/highlight.ts
@@ -4,7 +4,7 @@ import { Decoration, DecorationSet, EditorView, PluginValue, ViewPlugin, ViewUpd
 import { logger } from '@sourcegraph/common'
 import { Occurrence, SyntaxKind } from '@sourcegraph/shared/src/codeintel/scip'
 
-import { BlobInfo } from '../Blob'
+import type { BlobInfo } from '../CodeMirrorBlob'
 
 import { positionToOffset } from './utils'
